requirements.txt
Django>=4.2.4
gunicorn>=21.2.0
django-environ>=0.10.0
django-db-logger>=0.1.12
djangorestframework>=3.14.0
django-oauth-toolkit>=2.3.0 
Markdown>=3.4.4
django-filter>=23.2
django-cors-headers>=4.2.0
mysqlclient>=2.2.0
psycopg2>=2.9.7